+++ fractionalPart.h
@@ -2,8 +2,8 @@
 
 #define FRACTIONALPART_H_INCLUDED
 
-float *divisionMethodFractional(char *number, int *base, short int *detailed);
+float *divisionMethodFractional(char *number, int *base, int *detailed);
 
-char *multiplicationMethodFractional(char *number, int *base, short int *detailed);
+char *multiplicationMethodFractional(char *number, int *base, int *detailed);
 
 #endif // FRACTIONALPART_H_INCLUDED
