+++ integerPart.h
@@ -4,13 +4,13 @@
 
 int *transformNumberToDecimal (char *characterToTransform);
 
-int *multiplicationMethodInteger(char *number, int *base, short int *detailed);
+int *multiplicationMethodInteger(char *number, int *base, int *detailed);
 
 void setEquivalentDigit(char *currentDigit, int *reminder);
 
 void reverseString (char *stringToReverse);
 
-char *divisionMethodInteger(char *number, int *base, short int *detailed);
+char *divisionMethodInteger(char *number, int *base, int *detailed);
 
 
 #endif // INTEGERPART_H_INCLUDED
